% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_pcornet.R
\name{build_condition_era}
\alias{build_condition_era}
\title{Synthesize Condition Eras}
\usage{
build_condition_era(con)
}
\arguments{
\item{con}{Active DuckDB connection}
}
\description{
Synthesize Condition Eras
}
