% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_pcornet.R
\name{build_observation_period}
\alias{build_observation_period}
\title{Synthesize Observation Period envelopes}
\usage{
build_observation_period(con)
}
\arguments{
\item{con}{Active DuckDB connection}
}
\description{
Synthesize Observation Period envelopes
}
