% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{DEFAULT_MEDICATION_CONCEPTS}
\alias{DEFAULT_MEDICATION_CONCEPTS}
\title{Default Core Chronic Medication Classes (ATC / RxNorm Ancestors)}
\usage{
DEFAULT_MEDICATION_CONCEPTS
}
\description{
Default Core Chronic Medication Classes (ATC / RxNorm Ancestors)
}
