% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{list_available_phenotypes}
\alias{list_available_phenotypes}
\title{List available pre-compiled phenotype bundle names}
\usage{
list_available_phenotypes()
}
\value{
Character vector of phenotype names.
}
\description{
List available pre-compiled phenotype bundle names
}
