% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_pcornet.R
\name{load_cdm_source}
\alias{load_cdm_source}
\title{Populate CDM_SOURCE metadata table}
\usage{
load_cdm_source(
  con,
  cdm_source_name = NULL,
  cdm_holder = NULL,
  site_id = NULL,
  site_anon = NULL,
  site_name = NULL
)
}
\arguments{
\item{con}{Active DuckDB connection}

\item{cdm_source_name}{Optional CDM source name}

\item{cdm_holder}{Optional CDM data holder}

\item{site_id}{Optional numeric site identifier}

\item{site_anon}{Optional publication pseudonym (e.g. "Site A")}

\item{site_name}{Optional institutional name}
}
\description{
Populate CDM_SOURCE metadata table
}
