% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_pcornet.R
\name{load_location}
\alias{load_location}
\title{Maps \code{lds_address_history.csv} into OMOP CDM \code{location} table and updates \code{person.location_id}}
\usage{
load_location(con, source_dir, site_id = NULL, disambiguate_patids = FALSE)
}
\arguments{
\item{con}{Active DuckDB DBI connection.}

\item{source_dir}{Directory containing PCORnet CSV extracts.}

\item{site_id}{Optional site ID.}

\item{disambiguate_patids}{Logical.}
}
\value{
Invisibly NULL.
}
\description{
Maps \code{lds_address_history.csv} into OMOP CDM \code{location} table and updates \code{person.location_id}
}
