% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\docType{class}
\name{phenotypes}
\alias{phenotypes}
\title{Pre-compiled OHDSI Phenotype Library Concept Bundles}
\description{
Provides zero-dependency access to curated standard concept sets, ICD-9/10 source codes,
and domain classifications for major clinical phenotypes.
}
