% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remapping.R
\name{remap_all}
\alias{remap_all}
\title{Remap all supported clinical tables and optionally rebuild eras}
\usage{
remap_all(con, dry_run = FALSE, rebuild_eras = TRUE)
}
\arguments{
\item{con}{Active DuckDB connection}

\item{dry_run}{If TRUE, calculates and returns statistics without modifying tables}

\item{rebuild_eras}{Rebuild condition_era and drug_era if tables are modified}
}
\value{
Named list of remapping results per table
}
\description{
Remap all supported clinical tables and optionally rebuild eras
}
