% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{to_latex}
\alias{to_latex}
\title{Convert Object to LaTeX Table}
\usage{
to_latex(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character string containing the LaTeX code.
}
\description{
Convert Object to LaTeX Table
}
