% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{to_markdown}
\alias{to_markdown}
\title{Convert Object to Markdown Table}
\usage{
to_markdown(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character string containing the Markdown table.
}
\description{
Convert Object to Markdown Table
}
