% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{to_mermaid}
\alias{to_mermaid}
\title{Convert Object to Mermaid Diagram}
\usage{
to_mermaid(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character string containing the Mermaid diagram definition.
}
\description{
Convert Object to Mermaid Diagram
}
